vdev Core.Rpc

This module contains functions that are called from other pods via RPC.

Types

person()

person() :: %Core.Person{
  gender: binary(),
  inserted_by: binary(),
  documents: [map()],
  tax_id: binary(),
  second_name: binary(),
  email: binary(),
  is_active: boolean(),
  birth_country: binary(),
  first_name: binary(),
  death_date: Date,
  patient_signed: boolean(),
  preferred_way_communication: binary(),
  __meta__: term(),
  process_disclosure_data_consent: boolean(),
  last_name: binary(),
  phones: [map()],
  id: binary(),
  confidant_person: map(),
  birth_date: Date,
  version: binary(),
  addresses: [map()],
  unzr: binary(),
  secret: binary(),
  status: binary(),
  merged_ids: [binary()],
  birth_settlement: binary(),
  invalid_tax_id: boolean(),
  inserted_at: DateTime,
  emergency_contact: map(),
  updated_at: DateTime,
  authentication_methods: map(),
  no_tax_id: boolean(),
  updated_by: binary()
}

Functions

search_persons(params)

search_persons(params :: map()) ::
  {:error, any()}
  | %Scrivener.Page{
      entries: [person()],
      page_number: number(),
      page_size: number(),
      total_entries: number(),
      total_pages: number()
    }

Searches person by given parameters

Available parameters:

Parameter	Type	Example	Description

first_name	binary	Петро

last_name	binary	Іванов

second_name	binary	Миколайович

birth_date	DateTime	1991-08-19T00:00:00.000Z

tax_id	binary	3126509816

unzr	binary	19900101-0001	The record number in the demographic register

phone_number	binary	+380508887700	Any phone number that is present in phones

auth_phone_number	binary	+380508887700	Phone number that is used for authentication

ids	list(binary)	[‘56615208-b41f-4f32-9221-d22eb2db7dfb’, …]	List of id’s

documents	list(map)	[%{“type” => “PASSPORT”, “number” => “123456”}, …]	List of documents to search by with or condition

Returns %Scrivener.Page{entries: list(person), page_number: number(), page_size: number(), total_entries: number(), total_pages: number()}.

Examples

iex> Core.Rpc.search_persons(%{"last_name" => "Іванов"})
%Scrivener.Page{
  entries: [
    %Core.Person{
      id: "26e673e1-1d68-413e-b96c-407b45d9f572",
      first_name: "Петро",
      last_name: "Іванов",
      second_name: "Миколайович",
      birth_country: "Ukraine",
      inserted_at: #DateTime<2018-11-05 09:54:49.406483Z>,
      updated_at: #DateTime<2018-11-05 09:54:49.406486Z>,
      inserted_by: "bc11085d-3e47-49a5-90c4-c787ee41c1ba",
      updated_by: "d221d7f1-81cb-44d3-b6d4-8d7e42f97ff9",
      invalid_tax_id: false,
      email: "petroivanov@email.com",
      unzr: "19900101-0001",
      process_disclosure_data_consent: true,
      addresses: [],
      death_date: nil,
      birth_settlement: "Київ",
      preferred_way_communication: nil,
      merged_ids: [],
      documents: [],
      gender: "MALE",
      patient_signed: true,
      no_tax_id: false,
      secret: "sEcReT",
      birth_date: ~D[1991-08-27],
      is_active: true,
      version: "1",
      emergency_contact: %{},
      tax_id: "3126509816",
      confidant_person: nil,
      authentication_methods: [
        %{"phone_number" => "+380508887700", "type" => "OTP"}
      ],
      phones: [],
      status: "active"
    }
  ],
  page_number: 1,
  page_size: 1,
  total_entries: 1,
  total_pages: 1
}