v2.4.3 MPI.Rpc

This module contains functions that are called from other pods via RPC.

Types

merge_candidate()

merge_candidate() :: %{
  status: binary(),
  config: map(),
  details: map(),
  score: float(),
  person: person(),
  master_person: person(),
  inserted_at: DateTime,
  updated_at: DateTime
}

person()

person() :: %{
  process_disclosure_data_consent: boolean(),
  death_date: Date,
  documents: [map()],
  tax_id: binary(),
  second_name: binary(),
  secret: binary(),
  birth_country: binary(),
  email: binary(),
  first_name: binary(),
  no_tax_id: boolean(),
  invalid_tax_id: boolean(),
  preferred_way_communication: binary(),
  inserted_by: binary(),
  last_name: binary(),
  phones: [map()],
  id: binary(),
  inserted_at: DateTime,
  birth_date: Date,
  version: binary(),
  patient_signed: boolean(),
  unzr: binary(),
  emergency_contact: map(),
  updated_at: DateTime,
  status: binary(),
  addresses: [map()],
  merged_persons: [map()],
  master_person: %{},
  confidant_person: map(),
  is_active: boolean(),
  updated_by: binary(),
  authentication_methods: [map()],
  gender: binary(),
  birth_settlement: binary()
}

successfull_search_response()

successfull_search_response() :: %Scrivener.Page{
  entries: [person()],
  page_number: number(),
  page_size: number(),
  total_entries: number(),
  total_pages: number()
}

Functions

create_or_update_person(params, consumer_id)

create_or_update_person(map(), binary()) ::
  {:ok, person()} | Ecto.Changeset.t() | nil | {:error, any()}

Creates or updates person

Available parameters

Parameter	Type	Example	Description

id	binary	“dfa49f43-a4e6-4963-87ee-7e3aca1f3731”	Primary key identifier from the database

version	binary	“default”	Record version

first_name	binary	“Петро”	Patient first name

last_name	binary	“Іванов”	Patient last name

second_name	binary	“Миколайович”	Patient second name

birth_date	Date.t	~D[2000-08-19]	Patient birth date

birth_country	binary	“Україна”	Patient birth country

birth_settlement	binary	“Вінниця”	Patient birth settlement

gender	binary	“MALE”	Patient gender

email	binary	“qq2234562qq@gmail.com”	Patient’s email

tax_id	binary	“3378115538”	National person identifier

unzr	binary	“19900101-0001”	The unique number in Unified State Register

death_date	Date.t	~D[2000-08-19]	Patient death date

preferred_way_communication	binary	“email”	The way how a patient wants to be reached

invalid_tax_id	boolean	false	Flag to show if person has invalid tax_id

is_active	boolean	true	Flag to show whether person is active in system

secret	binary	“secret”	Person secret word

emergency_contact	map	%{“first_name” => “Петро”, “last_name” => “Іванов”, …}	Patient’s contract person in case of emergency

confidant_person	map	%{“birth_country” => “Україна”, …}	The person(s) who is(are) responsible for the patient

patient_signed	boolean	true	Flag to show that patient has signed declaration

process_disclosure_data_consent	boolean	true	Flag to show that person allowed to read personal data

status	binary	active	Patient status in the system

authentication_methods	list(map)	[%{“phone_number” => “+380955947998”, “type” => “OTP”}]	The method to verify changes of patient by patient

no_tax_id	boolean	false	Flag to show whether person rejected to have taxId

addresses	list(map)	[%{“apartment” => “23”, “area” => “ЛЬВІВСЬКА”, … }]	Patient addresses

phones	list(map)	[%{“number” => “+380955947998”, “type” => “MOBILE”}]	Patient phones

documents	list(map)	[%{“number” => “120518”, “type” => “PASSPORT”}]	Patient identification documents

Examples

iex> MPI.Rpc.create_or_update_person(%{“id” => “6e8d4595-e83c-4f97-be76-c6e2b96b05f1”, “birth_date” => “1990-01-01”}, “26e673e1-1d68-413e-b96c-407b45d9f572”) {

:ok,
%{
  id: "6e8d4595-e83c-4f97-be76-c6e2b96b05f1",
  authentication_methods: [
    %{"type" => "NA"}
  ],
  ...
}

}

get_auth_method(id)

get_auth_method(id :: binary()) :: nil | {:ok, map()}

Get authentication method for person by id

Examples

iex> MPI.Rpc.get_auth_method("26e673e1-1d68-413e-b96c-407b45d9f572")
{:ok, %{"type" => "OTP", "phone_number" => "+380630000000"}}

get_person_by_id(id, fields \\ nil)

get_person_by_id(id :: binary(), fields :: list() | nil) ::
  nil | {:ok, person()}

Resolves person by id Optionally, a list of fields could be given to limit the selection of fields for the response

Examples

iex> MPI.Rpc.get_person_by_id("26e673e1-1d68-413e-b96c-407b45d9f572")
{
  :ok,
  %{
    id: "26e673e1-1d68-413e-b96c-407b45d9f572",
    first_name: "Петро",
    last_name: "Іванов",
    second_name: "Миколайович",
    birth_country: "Ukraine",
    inserted_at: #DateTime<2018-11-05 09:54:49.406483Z>,
    updated_at: #DateTime<2018-11-05 09:54:49.406486Z>,
    inserted_by: "bc11085d-3e47-49a5-90c4-c787ee41c1ba",
    updated_by: "d221d7f1-81cb-44d3-b6d4-8d7e42f97ff9",
    invalid_tax_id: false,
    email: "petroivanov@email.com",
    unzr: "19900101-0001",
    process_disclosure_data_consent: true,
    addresses: [%{
      "APARTMENT" => "",
      "BUILDING" => "48а",
      "COUNTRY" => "UA",
      "SETTLEMENT" => "ХАРКІВ",
      "SETTLEMENT_ID" => "1241d1f9-ae81-4fe5-b614-f4f780a5acf0",
      "SETTLEMENT_TYPE" => "CITY",
      "STREET" => "Героїв Праці",
      "STREET_TYPE" => " STREET",
      "TYPE" => "REGISTRATION",
      "area" => "ХАРКІВСЬКА"
    }],
    death_date: nil,
    birth_settlement: "Київ",
    preferred_way_communication: nil,
    merged_persons: [],
    master_person: nil,
    documents: [],
    gender: "MALE",
    patient_signed: true,
    no_tax_id: false,
    secret: "sEcReT",
    birth_date: ~D[1991-08-27],
    is_active: true,
    version: "1",
    emergency_contact: %{},
    tax_id: "3126509816",
    confidant_person: nil,
    authentication_methods: [
      %{"phone_number" => "+380508887700", "type" => "OTP"}
    ],
    phones: [],
    status: "active"
  }
}

iex> MPI.Rpc.get_person_by_id("26e673e1-1d68-413e-b96c-407b45d9f572", [:id, :first_name, :last_name, :second_name, :birth_date])
{
  :ok,
  %{
    birth_date: ~D[1991-08-27],
    first_name: "Петро",
    id: "26e673e1-1d68-413e-b96c-407b45d9f572",
    last_name: "Іванов",
    second_name: "Миколайович"
  }
}

ql_search(filter, order_by \\ [], cursor \\ nil)

List person by filtering params (GQL only)

Examples

iex> MPI.Rpc.ql_search([{:last_name, :equal, "Іванов"}], [asc: :inserted_at], {100, 50})
{:ok, [
    %{
      id: "26e673e1-1d68-413e-b96c-407b45d9f572",
      first_name: "Петро",
      last_name: "Іванов",
      second_name: "Миколайович",
      birth_country: "Ukraine",
      inserted_at: #DateTime<2018-11-05 09:54:49.406483Z>,
      updated_at: #DateTime<2018-11-05 09:54:49.406486Z>,
      inserted_by: "bc11085d-3e47-49a5-90c4-c787ee41c1ba",
      updated_by: "d221d7f1-81cb-44d3-b6d4-8d7e42f97ff9",
      invalid_tax_id: false,
      email: "petroivanov@email.com",
      unzr: "19900101-0001",
      process_disclosure_data_consent: true,
      addresses: [],
      death_date: nil,
      birth_settlement: "Київ",
      preferred_way_communication: nil,
      merged_persons: [],
      master_person: nil,
      documents: [],
      gender: "MALE",
      patient_signed: true,
      no_tax_id: false,
      secret: "sEcReT",
      birth_date: ~D[1991-08-27],
      is_active: true,
      version: "1",
      emergency_contact: %{},
      tax_id: "3126509816",
      confidant_person: nil,
      authentication_methods: [
        %{"phone_number" => "+380508887700", "type" => "OTP"}
      ],
      phones: [],
      status: "active"
    }
  ]
}

reset_auth_method(id, actor_id)

reset_auth_method(id :: binary(), actor_id :: binary()) ::
  {:error, term()} | {:ok, person()}

Reset authentication method for person by id with actor_id UUID

Examples

iex> MPI.Rpc.reset_auth_method("26e673e1-1d68-413e-b96c-407b45d9f572", "22f673e1-1d68-413e-b96c-407b45d9ffa3")
{
  :ok,
  %{
    id: "26e673e1-1d68-413e-b96c-407b45d9f572",
    authentication_methods: [
      %{"type" => "NA"}
    ],
    ...
  }
}

search_persons(params, fields \\ nil, options \\ [])

search_persons(params :: map(), fields :: list() | nil, options :: list()) ::
  {:error, any()} | {:ok, [person()]}

Searches person by given parameters and options

Available parameters:

Parameter	Type	Example	Description

first_name	binary	Петро

last_name	binary	Іванов

second_name	binary	Миколайович

birth_date	DateTime	1991-08-19T00:00:00.000Z

tax_id	binary	3126509816

unzr	binary	19900101-0001	The record number in the demographic register

phone_number	binary	+380508887700	Any phone number that is present in phones

auth_phone_number	binary	+380508887700	Phone number that is used for authentication

ids	list(binary)	[‘56615208-b41f-4f32-9221-d22eb2db7dfb’, …]	List of id’s

documents	list(map)	[%{“type” => “PASSPORT”, “number” => “123456”}, …]	List of documents to search by with or condition

Returns %Scrivener.Page{entries: list(person), page_number: number(), page_size: number(), total_entries: number(), total_pages: number()}.

Available options:

read_only :: true | false. Default nil equal to false. If true, read records from read db only (replica)

paginate :: true | false. Default nil equal to false. If true, do Repo.paginate, else do Repo.all with limit, offset

Examples

iex> MPI.Rpc.search_persons(%{"last_name" => "Іванов"})
%Scrivener.Page{
  entries: [
    %{
      id: "26e673e1-1d68-413e-b96c-407b45d9f572",
      first_name: "Петро",
      last_name: "Іванов",
      second_name: "Миколайович",
      birth_country: "Ukraine",
      inserted_at: #DateTime<2018-11-05 09:54:49.406483Z>,
      updated_at: #DateTime<2018-11-05 09:54:49.406486Z>,
      inserted_by: "bc11085d-3e47-49a5-90c4-c787ee41c1ba",
      updated_by: "d221d7f1-81cb-44d3-b6d4-8d7e42f97ff9",
      invalid_tax_id: false,
      email: "petroivanov@email.com",
      unzr: "19900101-0001",
      process_disclosure_data_consent: true,
      addresses: [],
      death_date: nil,
      birth_settlement: "Київ",
      preferred_way_communication: nil,
      master_person: nil,
      merged_persons: [],
      documents: [],
      gender: "MALE",
      patient_signed: true,
      no_tax_id: false,
      secret: "sEcReT",
      birth_date: ~D[1991-08-27],
      is_active: true,
      version: "1",
      emergency_contact: %{},
      tax_id: "3126509816",
      confidant_person: nil,
      authentication_methods: [
        %{"phone_number" => "+380508887700", "type" => "OTP"}
      ],
      phones: [],
      status: "active"
    }
  ],
  page_number: 1,
  page_size: 1,
  total_entries: 1,
  total_pages: 1
}

iex> MPI.Rpc.search_persons(%{"ids" => ["26e673e1-1d68-413e-b96c-407b45d9f572"]}, [:id, :first_name, :last_name], read_only: true, paginate: true)
{:ok, [
    %{
      id: "26e673e1-1d68-413e-b96c-407b45d9f572",
      first_name: "Петро",
      last_name: "Іванов"
    }
  ]
}